fix: moved extract _language_code to llm service
