+++ src/bot/conversation.py
@@ -4,7 +4,7 @@ This file contains the functions to be called by the bot that are used to start 
 from src.repository.vocabulary import get_all_words, get_or_create_user
 from services.llm_service import llm_response, translate_to_language
 from services.language_service import language_name_from_code
-from src.bot.conversation_handling import send_bot_response
+from bot.helpers import send_bot_response
 
 def initializeConversation(message, bot):
     user = get_or_create_user(str(message.from_user.id), message.from_user.username, message.from_user.first_name, message.from_user.last_name)
