+++ src/services/llm_service.py
@@ -61,3 +61,26 @@ def translate_to_language(base_language_code, text, additional_system_message = 
     else:
         sentence_in_base_language = translate_sentence_with_llm(text, base_language_code, additional_system_message)
     return sentence_in_base_language
+
+
+def extracat_language_code_with_llm(user_input: str) -> str | None:
+    """
+    Function to extract the language code from the user message. the user message should be the name of the language they want to learn, e.g. spanish, arabo, italiano, frances, etc. 
+    given the user input the llm should respond either with a language code in format ISO 639-1 or ISO 639-2, e.g. 'es' for spanish, 'ar' for arabic, 'it' for italian, 'fr' for french, etc., or respond  with error message, informing the user tbat tbere was a mistake
+    then create an if else. if the result from the llm is a language code, then you can go on with the conversation and ask the user for the base language. if the result is an error message, then you should ask the user to repeat the language they want to learn
+        
+    args:
+    user_message: the message object from the user
+
+    returns:
+    language_code: the language code of the language from user_input
+
+    """
+    system_message = "We ask information about what language the user in interested in, and given the user input, you will have to extract in lower case the language code of the language selected by the user (NOT THE LANGUAGE CODE THAT THE USER IS TYPING IN!!). For example, if the user says 'English', you have to respond 'en', if the user says 'inglese', you have to respond 'en', if the user says 'spagnolo', you have to respond 'es'. In case the user message is not containing infromation around a language (for example they wrote a random word or sentence unrelated), respond with 'None'"
+    messages = []
+    messages.append({"role": "system", "content": system_message})
+    messages.append({"role": "user", "content": user_input})
+    learning_language_code = llm_response(messages)
+    if learning_language_code == "None":
+        return None
+    return learning_language_code
